Try to run solr as current user
- ensures data dir is owned by current user
- mount log dir outside

We could try to use a Docker volume for this, but the problem is that it
still wouldn't be owned by the solr user by default. If we were using a
Dockerfile instead of mounting config directories in, we would have some
other options. There might be other ways to work around this in the
future, but this works for now.