Remove nginx
- rename apache-cgi to apache
- add catalog rewriterules to apache config
- adjust proxying to proxy to catalog (via php fastcgi)
- add proxying to imgsrv fastcgi

This is a bit of a departure for how things work in production,
primarily because 1) we're using php via fastcgi, not via mod_php; 2)
everything is running under a single virtual host, which requires some
hackery to get working. (n.b. If we also wanted to add wordpress, we'd
need to be more explicit about which paths should be handled by the
catalog..)

Not yet tested: does auth work as expected for imgsrv?